packages/curate-worker/pyproject.toml
[project]
name = "curate-worker"
version = "0.1.0"
description = "Agent pipeline worker for Curate."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "curate-common",
    "agent-framework-core>=1.0.0rc1",
    "httpx>=0.28.1",
    "azure-servicebus>=7.14.0",
    "azure-ai-projects>=2.0.0b3",
    "azure-monitor-opentelemetry>=1.8.6",
]

[project.scripts]
curate-worker = "curate_worker.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/curate_worker"]
